Implement assigning requests to duckiebots based on a breadth first search
When there are pending customer requests, the taxi_central_node does a breadth first search
on the graph starting from the customers location to identify the closest duckiebot. If there
is one, it is assigned the request. If there are non available the request will have to wait.

This implementation is untested.